+++ src/components/OutlineListItem.vue
@@ -14,6 +14,14 @@
       </ul>
       <q-item-tile sublabel lines="3" v-if="outline.notes !== '' && showNotes">{{ outline.notes }}</q-item-tile>
       <br/>
+      <span v-if="this.$selectedTopic.get()">
+        <span v-if="showTopicAdd()">
+          <q-chip icon="fa-plus" @click.native="topicAdd" class="cursor-pointer" color="primary" small>Add</q-chip>&nbsp;
+        </span>
+        <span v-if="!showTopicAdd()">
+          <q-chip icon="fa-check" @click.native="topicAdd" color="positive" small>Added!</q-chip>&nbsp;
+        </span>
+      </span>
       <span>
         <q-icon name="fa-pencil" @click.native="openEdit" class="cursor-pointer" />&nbsp;
       </span>
@@ -82,6 +90,19 @@ export default {
       console.log('fin', value)
       this.outline = value
       this.editOpen = false
+    },
+    showTopicAdd () {
+      return this.$selectedTopic.get() && !this.$selectedTopic.find(this.outline._id)
+    },
+    topicAdd () {
+      var obj = {
+        topic: this.$selectedTopic.get().id,
+        media: this.outline._id,
+        type: 'outline'
+      }
+      this.database.add('resource', obj, (res) => {
+        this.$selectedTopic.add(this.outline._id)
+      })
     }
   }
 }